Bug fix: obtain_lock() and error logging...
If obtain_lock() failed to obtain the lock on a file (maybe due to lack of permissions), obtain_lock() would correctly return an array from handle_error().

However, methods calling obtain_lock() did not handle this array properly, and this resulted in a infinite loop in the http_stream_file() method. This problem is now fixed, and future errors on fread is not supressed with "@" just to be safe.