Migrated to Exception throwing...
The File Handler will now throw an Exception object instead of returning an array. Some of the array data is still available via $e->getExceptionArray();

This update should not break existing code, but checks such as "if(is_array())..." is now without function and can be removed. Instead, we can now use try and catch blocks to handle errors from the File Handler.